+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.9.0
-* Rewrote pretty much everything.
+* Rewrote pretty much everything to add downsampling.
 
 # 0.8.4
 * Add support for converting CCP4 mode 0 maps.